Move CopyForwardSchemeTask to its own header file
