+++ runtime/gc_vlhgc/CMakeLists.txt
@@ -1,5 +1,5 @@
 ################################################################################
-# Copyright (c) 2017, 2018 IBM Corp. and others
+# Copyright (c) 2017, 2019 IBM Corp. and others
 #
 # This program and the accompanying materials are made available under
 # the terms of the Eclipse Public License 2.0 which accompanies this
@@ -34,6 +34,7 @@ add_library(j9gcvlhgc STATIC
 	CopyForwardGMPCardCleaner.cpp
 	CopyForwardNoGMPCardCleaner.cpp
 	CopyForwardScheme.cpp
+	CopyForwardSchemeTask.cpp
 	CopyScanCacheChunkVLHGC.cpp
 	CopyScanCacheChunkVLHGCInHeap.cpp
 	CopyScanCacheListVLHGC.cpp
