+++ runtime/gc_vlhgc/CopyForwardScheme.cpp
@@ -50,6 +50,7 @@
 #include "ClassLoaderClassesIterator.hpp"
 #include "ClassLoaderIterator.hpp"
 #include "ClassLoaderRememberedSet.hpp"
+#include "CopyForwardSchemeTask.hpp"
 #include "CompactGroupManager.hpp"
 #include "CompactGroupPersistentStats.hpp"
 #include "CompressedCardTable.hpp"
@@ -126,170 +127,6 @@
  * if the common context disappears or becomes defined in a more complicated fashion
  */
 #define COMMON_CONTEXT_INDEX 0
-
-class MM_CopyForwardSchemeTask : public MM_ParallelTask
-{
-private:
-	MM_CopyForwardScheme *_copyForwardScheme;  /**< Tasks controlling scheme instance */
-	MM_CycleState *_cycleState;  /**< Collection cycle state active for the task */
-
-public:
-	virtual UDATA getVMStateID() { return OMRVMSTATE_GC_SCAVENGE; };
-
-	virtual void run(MM_EnvironmentBase *envBase)
-	{
-		MM_EnvironmentVLHGC *env = MM_EnvironmentVLHGC::getEnvironment(envBase);
-		_copyForwardScheme->workThreadGarbageCollect(env);
-	}
-
-	void setup(MM_EnvironmentBase *envBase)
-	{
-		MM_EnvironmentVLHGC *env = MM_EnvironmentVLHGC::getEnvironment(envBase);
-		if (env->isMasterThread()) {
-			Assert_MM_true(_cycleState == env->_cycleState);
-		} else {
-			Assert_MM_true(NULL == env->_cycleState);
-			env->_cycleState = _cycleState;
-		}
-		
-		env->_workPacketStats.clear();
-		env->_copyForwardStats.clear();
-		
-		/* record that this thread is participating in this cycle */
-		env->_copyForwardStats._gcCount = MM_GCExtensions::getExtensions(env)->globalVLHGCStats.gcCount;
-		env->_workPacketStats._gcCount = MM_GCExtensions::getExtensions(env)->globalVLHGCStats.gcCount;
-	}
-
-	void cleanup(MM_EnvironmentBase *envBase)
-	{
-		MM_EnvironmentVLHGC *env = MM_EnvironmentVLHGC::getEnvironment(envBase);
-
-		if (env->isMasterThread()) {
-			Assert_MM_true(_cycleState == env->_cycleState);
-		} else {
-			env->_cycleState = NULL;
-		}
-
-		env->_lastOverflowedRsclWithReleasedBuffers = NULL;
-	}
-
-	void masterCleanup(MM_EnvironmentBase *envBase)
-	{
-		MM_EnvironmentVLHGC *env = MM_EnvironmentVLHGC::getEnvironment(envBase);
-
-		MM_GCExtensions::getExtensions(env)->interRegionRememberedSet->resetOverflowedList();
-	}
-
-#if defined(J9MODRON_TGC_PARALLEL_STATISTICS)
-	void synchronizeGCThreads(MM_EnvironmentBase *env, const char *id);
-	bool synchronizeGCThreadsAndReleaseMaster(MM_EnvironmentBase *env, const char *id);
-	bool synchronizeGCThreadsAndReleaseSingleThread(MM_EnvironmentBase *env, const char *id);
-	
-	bool synchronizeGCThreadsAndReleaseMasterForAbort(MM_EnvironmentBase *env, const char *id);
-	void synchronizeGCThreadsForMark(MM_EnvironmentBase *env, const char *id);
-	bool synchronizeGCThreadsAndReleaseMasterForMark(MM_EnvironmentBase *env, const char *id);
-	void synchronizeGCThreadsForInterRegionRememberedSet(MM_EnvironmentBase *env, const char *id);
-#endif /* J9MODRON_TGC_PARALLEL_STATISTICS */
-
-	/**
-	 * Create a CopyForwardSchemeTask object.
-	 */
-	MM_CopyForwardSchemeTask(MM_EnvironmentVLHGC *env, MM_Dispatcher *dispatcher, MM_CopyForwardScheme *copyForwardScheme, MM_CycleState *cycleState) :
-		MM_ParallelTask((MM_EnvironmentBase *)env, dispatcher)
-		, _copyForwardScheme(copyForwardScheme)
-		, _cycleState(cycleState)
-	{
-		_typeId = __FUNCTION__;
-	}
-};
-
-#if defined(J9MODRON_TGC_PARALLEL_STATISTICS)
-void
-MM_CopyForwardSchemeTask::synchronizeGCThreads(MM_EnvironmentBase *envBase, const char *id)
-{
-	PORT_ACCESS_FROM_ENVIRONMENT(envBase);
-	MM_EnvironmentVLHGC* env = MM_EnvironmentVLHGC::getEnvironment(envBase);
-	U_64 startTime = j9time_hires_clock();
-	MM_ParallelTask::synchronizeGCThreads(env, id);
-	U_64 endTime = j9time_hires_clock();
-	env->_copyForwardStats.addToSyncStallTime(startTime, endTime);
-}
-
-bool
-MM_CopyForwardSchemeTask::synchronizeGCThreadsAndReleaseMaster(MM_EnvironmentBase *envBase, const char *id)
-{
-	PORT_ACCESS_FROM_ENVIRONMENT(envBase);
-	MM_EnvironmentVLHGC* env = MM_EnvironmentVLHGC::getEnvironment(envBase);
-	U_64 startTime = j9time_hires_clock();
-	bool result = MM_ParallelTask::synchronizeGCThreadsAndReleaseMaster(env, id);
-	U_64 endTime = j9time_hires_clock();
-	env->_copyForwardStats.addToSyncStallTime(startTime, endTime);
-
-	return result;
-}
-
-bool
-MM_CopyForwardSchemeTask::synchronizeGCThreadsAndReleaseSingleThread(MM_EnvironmentBase *envBase, const char *id)
-{
-	PORT_ACCESS_FROM_ENVIRONMENT(envBase);
-	MM_EnvironmentVLHGC* env = MM_EnvironmentVLHGC::getEnvironment(envBase);
-	U_64 startTime = j9time_hires_clock();
-	bool result = MM_ParallelTask::synchronizeGCThreadsAndReleaseSingleThread(env, id);
-	U_64 endTime = j9time_hires_clock();
-	env->_copyForwardStats.addToSyncStallTime(startTime, endTime);
-
-	return result;
-}
-
-bool
-MM_CopyForwardSchemeTask::synchronizeGCThreadsAndReleaseMasterForAbort(MM_EnvironmentBase *envBase, const char *id)
-{
-	PORT_ACCESS_FROM_ENVIRONMENT(envBase);
-	MM_EnvironmentVLHGC* env = MM_EnvironmentVLHGC::getEnvironment(envBase);
-	U_64 startTime = j9time_hires_clock();
-	bool result = MM_ParallelTask::synchronizeGCThreadsAndReleaseMaster(env, id);
-	U_64 endTime = j9time_hires_clock();
-	env->_copyForwardStats.addToAbortStallTime(startTime, endTime);
-
-	return result;
-}
-
-void
-MM_CopyForwardSchemeTask::synchronizeGCThreadsForMark(MM_EnvironmentBase *envBase, const char *id)
-{
-	PORT_ACCESS_FROM_ENVIRONMENT(envBase);
-	MM_EnvironmentVLHGC* env = MM_EnvironmentVLHGC::getEnvironment(envBase);
-	U_64 startTime = j9time_hires_clock();
-	MM_ParallelTask::synchronizeGCThreads(env, id);
-	U_64 endTime = j9time_hires_clock();
-	env->_copyForwardStats.addToMarkStallTime(startTime, endTime);
-}
-
-bool
-MM_CopyForwardSchemeTask::synchronizeGCThreadsAndReleaseMasterForMark(MM_EnvironmentBase *envBase, const char *id)
-{
-	PORT_ACCESS_FROM_ENVIRONMENT(envBase);
-	MM_EnvironmentVLHGC* env = MM_EnvironmentVLHGC::getEnvironment(envBase);
-	U_64 startTime = j9time_hires_clock();
-	bool result = MM_ParallelTask::synchronizeGCThreadsAndReleaseMaster(env, id);
-	U_64 endTime = j9time_hires_clock();
-	env->_copyForwardStats.addToMarkStallTime(startTime, endTime);
-
-	return result;
-}
-
-void
-MM_CopyForwardSchemeTask::synchronizeGCThreadsForInterRegionRememberedSet(MM_EnvironmentBase *envBase, const char *id)
-{
-	PORT_ACCESS_FROM_ENVIRONMENT(envBase);
-	MM_EnvironmentVLHGC* env = MM_EnvironmentVLHGC::getEnvironment(envBase);
-	U_64 startTime = j9time_hires_clock();
-	MM_ParallelTask::synchronizeGCThreads(env, id);
-	U_64 endTime = j9time_hires_clock();
-	env->_copyForwardStats.addToInterRegionRememberedSetStallTime(startTime, endTime);
-}
-
-#endif /* J9MODRON_TGC_PARALLEL_STATISTICS */
 
 MM_CopyForwardScheme::MM_CopyForwardScheme(MM_EnvironmentVLHGC *env, MM_HeapRegionManager *manager)
 	: MM_BaseNonVirtual()
